LabelClassify/pyproject.toml
[tool.poetry]
name = "labelclassify"
version = "0.1.0"
description = ""
authors = ["Lee Junhee <junhee0110@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
